+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated dependencies;
 - Message interface methods to accept `TransactionOptionsDto` instead of `TransactionOptions`;
 - `AccountHandle::stop_participating` renamed to `AccountHandle::stop_voting`;
+- `AccountMethod::StopParticipating` renamed to `AccountMethod::StopVoting`;
 
 ## 1.0.0-rc.4 - 2022-12-23